+++ src/v8_py_frontend/object_manipulator.cc
@@ -1,12 +1,15 @@
 #include "object_manipulator.h"
 #include <v8-container.h>
 #include <v8-context.h>
+#include <v8-exception.h>
+#include <v8-function.h>
 #include <v8-isolate.h>
 #include <v8-local-handle.h>
 #include <v8-object.h>
 #include <v8-persistent-handle.h>
 #include <v8-primitive.h>
-#include <string>
+#include <cstdint>
+#include <vector>
 #include "binary_value.h"
 
 namespace MiniRacer {
@@ -42,14 +45,115 @@ auto ObjectManipulator::GetOwnPropertyNames(
   return bv_factory_->FromValue(local_context, names);
 }
 
-auto ObjectManipulator::KeyToValue(v8::Isolate* isolate, const std::string& key)
-    -> v8::Local<v8::Value> {
-  return v8::String::NewFromUtf8(isolate, key.c_str()).ToLocalChecked();
+auto ObjectManipulator::Get(v8::Isolate* isolate,
+                            v8::Persistent<v8::Value>* object,
+                            BinaryValue* key) -> BinaryValue::Ptr {
+  const v8::Isolate::Scope isolate_scope(isolate);
+  const v8::HandleScope handle_scope(isolate);
+  const v8::Local<v8::Object> local_object =
+      object->Get(isolate).As<v8::Object>();
+  const v8::Local<v8::Context> local_context = context_->Get(isolate);
+
+  const v8::Local<v8::Value> local_key =
+      BinaryValueFactory::ToValue(local_context, key);
+
+  if (!local_object->Has(local_context, local_key).ToChecked()) {
+    // Return null if no item.
+    return {};
+  }
+
+  const v8::Local<v8::Value> value =
+      local_object->Get(local_context, local_key).ToLocalChecked();
+
+  return bv_factory_->FromValue(local_context, value);
 }
 
-auto ObjectManipulator::KeyToValue(v8::Isolate* isolate,
-                                   double key) -> v8::Local<v8::Value> {
-  return v8::Number::New(isolate, key);
+void ObjectManipulator::Set(v8::Isolate* isolate,
+                            v8::Persistent<v8::Value>* object,
+                            BinaryValue* key,
+                            BinaryValue* val) {
+  const v8::Isolate::Scope isolate_scope(isolate);
+  const v8::HandleScope handle_scope(isolate);
+  const v8::Local<v8::Object> local_object =
+      object->Get(isolate).As<v8::Object>();
+  const v8::Local<v8::Context> local_context = context_->Get(isolate);
+
+  const v8::Local<v8::Value> local_key =
+      BinaryValueFactory::ToValue(local_context, key);
+
+  const v8::Local<v8::Value> local_value =
+      BinaryValueFactory::ToValue(local_context, val);
+
+  local_object->Set(local_context, local_key, local_value).ToChecked();
+}
+
+auto ObjectManipulator::Del(v8::Isolate* isolate,
+                            v8::Persistent<v8::Value>* object,
+                            BinaryValue* key) -> bool {
+  const v8::Isolate::Scope isolate_scope(isolate);
+  const v8::HandleScope handle_scope(isolate);
+  const v8::Local<v8::Object> local_object =
+      object->Get(isolate).As<v8::Object>();
+  const v8::Local<v8::Context> local_context = context_->Get(isolate);
+
+  const v8::Local<v8::Value> local_key =
+      BinaryValueFactory::ToValue(local_context, key);
+
+  if (!local_object->Has(local_context, local_key).ToChecked()) {
+    return false;
+  }
+
+  return local_object->Delete(local_context, local_key).ToChecked();
+}
+
+auto ObjectManipulator::Splice(v8::Isolate* isolate,
+                               v8::Persistent<v8::Value>* object,
+                               int32_t start,
+                               int32_t delete_count,
+                               BinaryValue* new_val) -> BinaryValue::Ptr {
+  const v8::Isolate::Scope isolate_scope(isolate);
+  const v8::HandleScope handle_scope(isolate);
+  const v8::Local<v8::Object> local_object =
+      object->Get(isolate).As<v8::Object>();
+  const v8::Local<v8::Context> local_context = context_->Get(isolate);
+
+  // Array.prototype.splice doesn't exist in C++ in V8. We have to find the JS
+  // function and call it:
+  const v8::Local<v8::String> splice_name =
+      v8::String::NewFromUtf8Literal(isolate, "splice");
+
+  v8::Local<v8::Value> splice_val;
+  if (!local_object->Get(local_context, splice_name).ToLocal(&splice_val)) {
+    return bv_factory_->FromString("no splice method on object",
+                                   type_execute_exception);
+  }
+
+  if (!splice_val->IsFunction()) {
+    return bv_factory_->FromString("splice method is not a function",
+                                   type_execute_exception);
+  }
+
+  const v8::Local<v8::Function> splice_func = splice_val.As<v8::Function>();
+
+  const v8::TryCatch trycatch(isolate);
+
+  std::vector<v8::Local<v8::Value>> argv = {
+      v8::Int32::New(isolate, start),
+      v8::Int32::New(isolate, delete_count),
+  };
+  if (new_val != nullptr) {
+    argv.push_back(BinaryValueFactory::ToValue(local_context, new_val));
+  }
+
+  v8::MaybeLocal<v8::Value> maybe_value = splice_func->Call(
+      local_context, local_object, static_cast<int>(argv.size()), argv.data());
+  if (maybe_value.IsEmpty()) {
+    return bv_factory_->FromExceptionMessage(local_context, trycatch.Message(),
+                                             trycatch.Exception(),
+                                             type_execute_exception);
+  }
+
+  return bv_factory_->FromValue(local_context, maybe_value.ToLocalChecked());
 }
 
 }  // end namespace MiniRacer